+++ CHANGELOG.md
@@ -13,6 +13,13 @@
 * Ruby 2.7.0
 * Rename `EveOnline::ESI::UniverseRegion#constellations` to `#constellation_ids`
 * Remove `constellations` from `EveOnline::ESI::UniverseRegion#as_json`
+* Add `EveOnline::ESI::Models::Stargate#destination_stargate_id`
+* Add `EveOnline::ESI::Models::Stargate#destination_system_id`
+* Add `destination_stargate_id` and `destination_system_id` to `EveOnline::ESI::Models::Stargate#as_json`
+* Remove `EveOnline::ESI::Models::StargateDestination`
+* Remove `EveOnline::ESI::UniverseStargate#destination`
+* Add `EveOnline::ESI::UniverseStargate#destination_stargate_id`
+* Add `EveOnline::ESI::UniverseStargate#destination_system_id`
 
 ## v0.30.0